ETL Google Cloud Functions/requirements.txt.txt
google-cloud-bigquery
pandas
featuretools
pyarrow
